app/requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
pymysql
requests
python-dotenv
cryptography
numpy
